requirements.txt
pyTelegramBotAPI==4.7.0
python-decouple==3.6